cc evolve

+++ plugins/fractalsubctrler.h
@@ -55,8 +55,6 @@ struct _FRACTaLSubController
   GstClockTime              last_executed;
 
   gint32                    bottleneck_point;
-//  gint32                    stalled_bytes;
-//  gint32                    keeping_point;
   gint32                    inflection_point;
 
   GstClockTime              approvement_interval;
@@ -68,7 +66,7 @@ struct _FRACTaLSubController
   gdouble                   est_rr;
 
 
-  gint32                    congested_drift;
+  guint64                   congested_drift;
   gdouble                   congested_lost;
 
   gint32                    est_capacity;
@@ -113,6 +111,8 @@ struct _FRACTaLSubController
   gpointer                  priv;
 
   SndTracker*               sndtracker;
+
+
 
 };
